requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioredis==2.0.1
aiosignal==1.4.0
aiosql==13.4
aiosqlite==0.21.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.9
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
python-decouple==3.8
redis==6.2.0
setuptools==80.9.0
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1